Move Scaling related handler place.
