+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/main/java/org/apache/shardingsphere/proxy/backend/text/distsql/ral/scaling/query/QueryableScalingRALBackendHandlerFactory.java
@@ -15,34 +15,38 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.proxy.backend.text.distsql.ral.update;
+package org.apache.shardingsphere.proxy.backend.text.distsql.ral.scaling.query;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
-import org.apache.shardingsphere.distsql.parser.statement.ral.UpdatableRALStatement;
-import org.apache.shardingsphere.infra.distsql.update.RALUpdater;
+import org.apache.shardingsphere.distsql.parser.statement.ral.QueryableRALStatement;
+import org.apache.shardingsphere.infra.distsql.query.DistSQLResultSet;
+import org.apache.shardingsphere.proxy.backend.session.ConnectionSession;
 import org.apache.shardingsphere.proxy.backend.text.TextProtocolBackendHandler;
 import org.apache.shardingsphere.spi.ShardingSphereServiceLoader;
+import org.apache.shardingsphere.spi.typed.TypedSPIRegistry;
+
+import java.util.Properties;
 
 /**
- *  Updatable scaling RAL backend handler factory.
+ * Queryable scaling RAL backend handler factory.
  */
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class UpdatableScalingRALBackendHandlerFactory {
+public final class QueryableScalingRALBackendHandlerFactory {
     
     static {
-        ShardingSphereServiceLoader.register(RALUpdater.class);
+        ShardingSphereServiceLoader.register(DistSQLResultSet.class);
     }
     
     /**
      * Create new instance of queryable RAL backend handler.
      * 
      * @param sqlStatement queryable RAL statement
+     * @param connectionSession connection session
      * @return queryable RAL backend handler
      */
-    public static TextProtocolBackendHandler newInstance(final UpdatableRALStatement sqlStatement) {
-        UpdatableScalingRALBackendHandler result = new UpdatableScalingRALBackendHandler();
-        result.setSqlStatement(sqlStatement);
-        return result;
+    public static TextProtocolBackendHandler newInstance(final QueryableRALStatement sqlStatement, final ConnectionSession connectionSession) {
+        DistSQLResultSet resultSet = TypedSPIRegistry.getRegisteredService(DistSQLResultSet.class, sqlStatement.getClass().getCanonicalName(), new Properties());
+        return new QueryableScalingRALBackendHandler(sqlStatement, connectionSession, resultSet);
     }
 }
